ocr_service/requirements.txt
fastapi
uvicorn
python-dotenv
pytesseract
Pillow